多加盟商：优化查询在线活动

+++ src/main/java/com/xiliulou/electricity/service/ShareActivityService.java
@@ -71,5 +71,5 @@ public interface ShareActivityService {
      */
     R<?> removeById(Long id);
     
-    ShareActivity queryOnlineActivity(Integer tenantId, Integer franchiseeId);
+    ShareActivity queryOnlineActivity(Integer tenantId, Long franchiseeId);
 }
